+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.3 - Unreleased
 
+### Added
+* Add `ensure_video` method to `CocoDataset`.
+* Add `ensure_track` method to `CocoDataset`.
+* Add `sort_values` method to `ObjectList1D` (and hence vectorized `Images` and `Annots` objects).
+
 
 ## Version 0.7.2 - Released 2023-10-04