+++ CHANGELOG.md
@@ -13,6 +13,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Add `ensure_track` method to `CocoDataset`.
 * Add `sort_values` method to `ObjectList1D` (and hence vectorized `Images` and `Annots` objects).
 
+### Fixed
+* kwcoco.Coerce with sqlview now works for .kwcoco.zip files.
+
 
 ## Version 0.7.2 - Released 2023-10-04